requirements.txt
appdirs==1.4.4
argcomplete==3.3.0
barcode==1.0.4
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
escpos==2.0.0
Flask==3.0.3
future==1.0.0
idna==3.7
ingate-python @ file:///D:/HUSNI/PROJECT/test/ingate_python-0.1.tar.gz#sha256=b802d2a899b3c2cd9cfd5efd616322b245805eaf818946605b9cc56a487f147a
itsdangerous==2.2.0
Jinja2==3.1.3
Levenshtein==0.25.1
MarkupSafe==2.1.5
pillow==10.3.0
psycopg2==2.9.9
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.15.1
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pyusb==1.2.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.8.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
viivakoodi==0.8.0
Werkzeug==3.0.2
